requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
cloudinary==1.19.1
cycler==0.10.0
Django==2.2.5
django-bootstrap4==1.0.1
django-cloudinary-storage==0.2.3
django-environ==0.4.5
django-widget-tweaks==1.4.5
idna==2.8
jaconv==0.2.4
kiwisolver==1.1.0
matplotlib==3.1.1
mock==3.0.5
numpy==1.17.2
opencv-python==4.1.1.26
pandas==0.25.1
Pillow==6.2.0
pyocr==0.7.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
xlrd==1.2.0
